[1662] Make course description the #show page
This makes the URLs make sense from a semantic standpoint.
