+++ app/views/courses/about.html.erb
@@ -1,7 +1,7 @@
 <%= content_for :page_title, "About this course – #{course.name_and_code}" %>
 
 <% content_for :before_content do %>
-  <%= govuk_back_link_to(description_provider_course_path(@provider.provider_code, course.course_code)) %>
+  <%= govuk_back_link_to(provider_course_path(@provider.provider_code, course.course_code)) %>
 <% end %>
 
 <%= render partial: 'courses/copy_content_warning',
@@ -85,7 +85,7 @@
       <%= f.submit "Save", class: "govuk-button", data: { qa: 'course__save' } %>
 
       <p class="govuk-body">
-        <%= link_to 'Cancel changes', description_provider_course_path(@provider.provider_code, course.course_code), class: "govuk-link" %>
+        <%= link_to 'Cancel changes', provider_course_path(@provider.provider_code, course.course_code), class: "govuk-link" %>
       </p>
     <% end %>
   </div>
