+++ app/views/courses/details.html.erb
@@ -1,0 +1,57 @@
+<%= content_for :page_title, "#{course.name_and_code} - Courses" %>
+
+<% content_for :before_content do %>
+  <nav class="govuk-breadcrumbs">
+    <ol class="govuk-breadcrumbs__list">
+      <% if @has_multiple_providers then %>
+        <li class="govuk-breadcrumbs__list-item">
+          <%= link_to "Organisations", providers_path, class: "govuk-breadcrumbs__link" %>
+        </li>
+      <% end %>
+      <li class="govuk-breadcrumbs__list-item">
+        <%= link_to @provider.provider_name, provider_path(code: @provider.provider_code), class: "govuk-breadcrumbs__link" %>
+      </li>
+      <li class="govuk-breadcrumbs__list-item">
+        <%= link_to "Courses", provider_courses_path(@provider.provider_code), class: "govuk-breadcrumbs__link" %>
+      </li>
+      <li class="govuk-breadcrumbs__list-item" aria-current="page">
+        <%= course.name %> (<%= course.course_code %>)
+      </li>
+    </ol>
+  </nav>
+<% end %>
+
+<h1 class="govuk-heading-xl">
+  <span class="govuk-caption-xl"><%= course.description %></span>
+  <%= course.name %> (<%= course.course_code %>)
+</h1>
+
+<div class="course-tabs govuk-tabs">
+  <h2 class="govuk-tabs__title">Contents</h2>
+
+  <ul class="govuk-tabs__list" role="tablist">
+    <li class="govuk-tabs__list-item" role="presentation">
+      <%=
+        link_to provider_course_path(@provider.provider_code, course.course_code),
+        class: "govuk-tabs__tab", role: 'tab', tabindex: '0', aria: { selected: false } do
+      %>
+        Description<br>
+        <span class="govuk-body-s govuk-!-font-weight-regular pointer-events-none">Content, fees and eligibility</span>
+      <% end %>
+    </li>
+
+    <li class="govuk-tabs__list-item" role="presentation">
+      <%=
+        link_to "#basic_details",
+        class: "govuk-tabs__tab govuk-tabs__tab--selected", id: "basic_details_tab", role: 'tab', tabindex: '-1', aria: { selected: true } do
+      %>
+        Basic details<br>
+        <span class="govuk-body-s govuk-!-font-weight-regular pointer-events-none">Locations, outcome, subject</span>
+      <% end %>
+    </li>
+  </ul>
+
+  <section class="govuk-tabs__panel" id="basic_details" role="tabpanel" aria-labelledby="basic_details_tab">
+    <%= render partial: 'courses/basic_details_tab' %>
+  </section>
+</div>
