+++ spec/features/courses/requirements_spec.rb
@@ -29,7 +29,7 @@ feature 'Course requirements', type: :feature do
       "/providers/#{provider.provider_code}/courses/#{course.course_code}",
       course.render, :patch, 200
     )
-    visit description_provider_course_path(provider.provider_code, course.course_code)
+    visit provider_course_path(provider.provider_code, course.course_code)
     click_on 'Requirements and eligibility'
 
     expect(current_path).to eq requirements_provider_course_path('AO', course.course_code)
@@ -60,7 +60,7 @@ feature 'Course requirements', type: :feature do
       'Your changes have been saved'
     )
 
-    expect(current_path).to eq description_provider_course_path('AO', course.course_code)
+    expect(current_path).to eq provider_course_path('AO', course.course_code)
   end
 
   scenario 'submitting with validation errors' do
