+++ spec/features/courses/show_spec.rb
@@ -1,152 +1,240 @@
 require 'rails_helper'
 
-feature 'Show course', type: :feature do
-  let(:provider) { jsonapi(:provider, accredited_body?: false, sites: [site1, site2]) }
-  let(:course) {
-    jsonapi :course,
-            qualifications: %w[qts pgce],
-            study_mode: 'full_time',
-            start_date: Time.zone.local(2019),
-            sites: [site1, site2],
+feature 'Course show', type: :feature do
+  let(:provider) { jsonapi(:provider, accredited_body?: false) }
+  let(:course_jsonapi) {
+    jsonapi(:course,
+            has_vacancies?: true,
+            open_for_applications?: true,
+            funding: 'fee',
+            sites: [site],
             provider: provider,
-            accrediting_provider: provider,
-            open_for_applications?: true
+            fee_uk_eu: 9250,
+            last_published_at: '2019-03-05T14:42:34Z')
   }
-  let(:site1) { jsonapi(:site, location_name: 'London') }
-  let(:site2) { jsonapi(:site, location_name: 'Manchester') }
-  let(:site_status1) do
-    jsonapi(:site_status, :full_time, site: site1, status: 'running')
+  let(:site) { jsonapi(:site) }
+  let(:site_status) do
+    jsonapi(:site_status, :full_time_and_part_time, site: site)
   end
-  let(:site_status2) do
-    jsonapi(:site_status, :part_time, site: site2, status: 'suspended')
-  end
-  let(:course_response) { course.render }
+  let(:course)          { course_jsonapi.to_resource }
+  let(:course_response) { course_jsonapi.render }
   before do
     stub_omniauth
     stub_api_v2_request(
-      "/providers/A0/courses/#{course.course_code}?include=sites,provider.sites,accrediting_provider",
+      "/providers/#{provider.provider_code}/courses/#{course.course_code}?include=sites,provider.sites,accrediting_provider",
       course_response
     )
+    visit provider_course_path(provider.provider_code, course.course_code)
   end
 
   let(:course_page) { PageObjects::Page::Organisations::Course.new }
 
-  scenario 'viewing the show courses page' do
-    visit "/organisations/A0/courses/#{course.course_code}"
+  describe 'with a fee paying course' do
+    scenario 'it shows the course show page' do
+      expect(course_page.caption).to have_content(
+        course.description
+      )
+      expect(course_page.title).to have_content(
+        "#{course.name} (#{course.course_code})"
+      )
+      expect(course_page.about).to have_content(
+        course.about_course
+      )
+      expect(course_page.interview_process).to have_content(
+        course.interview_process
+      )
+      expect(course_page.placements_info).to have_content(
+        course.how_school_placements_work
+      )
+      expect(course_page.length).to have_content(
+        course.course_length
+      )
+      expect(course_page.uk_fees).to have_content(
+        '£9,250'
+      )
+      expect(course_page.international_fees).to have_content(
+        course.fee_international
+      )
+      expect(course_page.fee_details).to have_content(
+        course.fee_details
+      )
+      expect(course_page.required_qualifications).to have_content(
+        course.required_qualifications
+      )
+      expect(course_page.personal_qualities).to have_content(
+        course.personal_qualities
+      )
+      expect(course_page.other_requirements).to have_content(
+        course.other_requirements
+      )
+      expect(course_page.last_published_at).to have_content(
+        'Last published: 5 March 2019'
+      )
+      expect(course_page).to_not have_preview_link
 
-    expect(course_page)
-      .to be_displayed(provider_code: 'A0', course_code: course.course_code)
-
-    expect(course_page.caption).to have_content(
-      course.description
-    )
-    expect(course_page.title).to have_content(
-      "#{course.name} (#{course.course_code})"
-    )
-    expect(course_page.subjects).to have_content(
-      course.subjects.sort.join('').to_s
-    )
-    expect(course_page.qualifications).to have_content(
-      'PGCE with QTS'
-    )
-    expect(course_page.study_mode).to have_content(
-      'Full time'
-    )
-    expect(course_page.start_date).to have_content(
-      'January 2019'
-    )
-    expect(course_page.name).to have_content(
-      course.name
-    )
-    expect(course_page.description).to have_content(
-      course.description
-    )
-    expect(course_page.course_code).to have_content(
-      course.course_code
-    )
-    expect(course_page.locations).to have_content(
-      site1.location_name
-    )
-    expect(course_page.locations).to have_content(
-      site2.location_name
-    )
-    expect(course_page.edit_locations_link).to have_content(
-      "Change location"
-    )
-    expect(course_page).not_to have_manage_provider_locations_link
-    expect { course_page.apprenticeship }.to raise_error(Capybara::ElementNotFound)
-    expect(course_page.funding).to have_content(
-      'Fee paying (no salary)'
-    )
-    expect(course_page.accredited_body).to have_content(
-      provider.provider_name
-    )
-    expect(course_page.application_status).to have_content(
-      'Open'
-    )
-    expect(course_page.is_send).to have_content(
-      'No'
-    )
-    expect(course_page.level).to have_content(
-      'Secondary'
-    )
-    expect(course_page).to have_entry_requirements
-  end
-
-  context 'when the provider only has one location' do
-    let(:provider) { jsonapi(:provider, accredited_body?: false, sites: [site1]) }
-    let(:course) {
-      jsonapi :course,
-              site_statuses: [site_status1],
-              provider: provider,
-              accrediting_provider: provider,
-              ucas_status: 'new'
-    }
-
-    scenario 'viewing the show courses page' do
-      visit "/organisations/A0/courses/#{course.course_code}"
-
-      expect(course_page).not_to have_edit_locations_link
-      expect(course_page.manage_provider_locations_link).to have_content(
-        "Manage all your locations"
+      expect(course_page).to have_link(
+        'About this course',
+        href: "/organisations/#{provider.provider_code}/courses/#{course.course_code}/about"
+      )
+      expect(course_page).to have_link(
+        'Course length and fees',
+        href: "/organisations/#{provider.provider_code}/courses/#{course.course_code}/fees"
+      )
+      expect(course_page).to have_link(
+        'Requirements and eligibility',
+        href: "/organisations/#{provider.provider_code}/courses/#{course.course_code}/requirements"
       )
     end
   end
 
-  context 'when the course is new and not running' do
-    let(:course) {
-      jsonapi :course,
-              sites: [site1, site2],
+  describe 'with a salaried course' do
+    let(:course_jsonapi) {
+      jsonapi(:course,
+              funding: 'salary',
+              sites: [site],
               provider: provider,
-              accrediting_provider: provider,
-              ucas_status: 'new'
+              accrediting_provider: provider)
     }
+    let(:course)          { course_jsonapi.to_resource }
+    let(:course_response) { course_jsonapi.render }
 
-    scenario 'viewing the show courses page' do
-      visit "/organisations/A0/courses/#{course.course_code}"
-
-      expect(course_page.locations).to have_content(
-        site1.location_name
+    scenario 'it shows the course show page' do
+      expect(course_page.caption).to have_content(
+        course.description
       )
-      expect(course_page.locations).to have_content(
-        site2.location_name
+      expect(course_page.title).to have_content(
+        "#{course.name} (#{course.course_code})"
       )
+      expect(course_page.about).to have_content(
+        course.about_course
+      )
+      expect(course_page.interview_process).to have_content(
+        course.interview_process
+      )
+      expect(course_page.placements_info).to have_content(
+        course.how_school_placements_work
+      )
+      expect(course_page.length).to have_content(
+        course.course_length
+      )
+      expect(course_page.salary).to have_content(
+        course.salary_details
+      )
+      expect(course_page.required_qualifications).to have_content(
+        course.required_qualifications
+      )
+      expect(course_page.personal_qualities).to have_content(
+        course.personal_qualities
+      )
+      expect(course_page.other_requirements).to have_content(
+        course.other_requirements
+      )
+      expect(course_page).to_not have_preview_link
     end
   end
 
-  scenario 'viewing the show page for a course that does not exist' do
-    stub_api_v2_request(
-      "/providers/ZZ/courses/ZZZ?include=sites,provider.sites,accrediting_provider",
-      '',
-      :get,
-      404
-    )
+  context 'when the course is running' do
+    let(:course_jsonapi) {
+      jsonapi(:course,
+              findable?: true,
+              content_status: 'published',
+              ucas_status: 'running',
+              has_vacancies?: true,
+              open_for_applications?: true,
+              provider: provider)
+    }
+    let(:course)          { course_jsonapi.to_resource }
+    let(:course_response) { course_jsonapi.render }
 
-    course
-    visit "/organisations/ZZ/courses/ZZZ"
+    scenario 'it displays a status panel' do
+      expect(course_page).to have_status_panel
+      expect(course_page.is_findable).to have_content('Yes')
+      expect(course_page.has_vacancies).to have_content('Yes')
+      expect(course_page.open_for_applications).to have_content('Open')
+      expect(course_page.status_tag).to have_content('Published')
+    end
+  end
 
-    expect(course_page)
-    .to be_displayed(provider_code: 'ZZ', course_code: 'ZZZ')
-    expect(course_page.title.text).to eq 'Page not found'
+  context 'when the course is not running' do
+    let(:course_jsonapi) {
+      jsonapi(:course,
+              ucas_status: 'not_running',
+              provider: provider)
+    }
+    let(:course)          { course_jsonapi.to_resource }
+    let(:course_response) { course_jsonapi.render }
+
+    scenario 'it hides the status panel' do
+      expect(course_page).not_to have_status_panel
+    end
+
+    scenario 'it shows a warning about the course status' do
+      expect(course_page).to have_content('This course is not running.')
+    end
+  end
+
+  context 'when the course is new' do
+    let(:course_jsonapi) {
+      jsonapi(:course,
+              findable?: false,
+              content_status: 'draft',
+              ucas_status: 'new',
+              provider: provider)
+    }
+    let(:course)          { course_jsonapi.to_resource }
+    let(:course_response) { course_jsonapi.render }
+
+    scenario 'it displays a status panel' do
+      expect(course_page).to have_status_panel
+      expect(course_page.is_findable).to have_content('No')
+      expect(course_page.status_tag).to have_content('Draft')
+      expect(course_page).to have_preview_link
+    end
+
+    describe 'publishing' do
+      before do
+        stub_api_v2_request(
+          "/providers/#{provider.provider_code}/courses/#{course.course_code}?include=sites,provider.sites,accrediting_provider",
+          course_response
+        )
+      end
+
+      context 'without errors' do
+        before do
+          stub_api_v2_request(
+            "/providers/#{provider.provider_code}/courses/#{course.course_code}/publish",
+            nil,
+            :post
+          )
+        end
+
+        scenario 'it shows the show page with success flash' do
+          course_page.publish.click
+
+          expect(course_page).to be_displayed
+          expect(course_page.success_summary).to have_content("Your course has been published.")
+          expect(course_page.success_summary).to have_content("The link for this course is: https://localhost:5000/course/#{provider.provider_code}/#{course.course_code}")
+        end
+      end
+
+      context 'with errors' do
+        before do
+          stub_api_v2_request(
+            "/providers/#{provider.provider_code}/courses/#{course.course_code}/publish",
+            build(:error, :for_course_publish),
+            :post,
+            422
+          )
+        end
+
+        scenario 'it shows the show page with validation errors' do
+          course_page.publish.click
+
+          expect(page.title).to have_content('Error:')
+          expect(course_page).to be_displayed
+          expect(course_page.error_summary).to have_content("About course can't be blank")
+        end
+      end
+    end
   end
 end
