+++ spec/site_prism/page_objects/page/organisations/course.rb
@@ -4,23 +4,28 @@ module PageObjects
       class Course < CourseBase
         set_url '/organisations/{provider_code}/courses/{course_code}'
 
-        element :qualifications, '[data-qa=course__qualifications]'
-        element :study_mode, '[data-qa=course__study_mode]'
-        element :start_date, '[data-qa=course__start_date]'
-        element :name, '[data-qa=course__name]'
-        element :description, '[data-qa=course__description]'
-        element :course_code, '[data-qa=course__course_code]'
-        element :locations, '[data-qa=course__locations]'
-        element :edit_locations_link, '[data-qa=course__edit_locations_link]'
-        element :manage_provider_locations_link, '[data-qa=course__manage_provider_locations_link]'
-        element :apprenticeship, '[data-qa=course__apprenticeship]'
-        element :funding, '[data-qa=course__funding]'
-        element :accredited_body, '[data-qa=course__accredited_body]'
-        element :application_status, '[data-qa=course__application_status]'
-        element :is_send, '[data-qa=course__is_send]'
-        element :subjects, '[data-qa=course__subjects]'
-        element :level, '[data-qa=course__level]'
-        element :entry_requirements, '[data-qa=course__entry_requirements]'
+        element :about, '[data-qa=course__about_course]'
+        element :interview_process, '[data-qa=course__interview_process]'
+        element :placements_info, '[data-qa=course__how_school_placements_work]'
+        element :length, '[data-qa=course__course_length]'
+        element :salary, '[data-qa=course__salary_details]'
+        element :uk_fees, '[data-qa=course__fee_uk_eu]'
+        element :international_fees, '[data-qa=course__international_fees]'
+        element :fee_details, '[data-qa=course__fee_details]'
+        element :financial_support, '[data-qa=course__financial_support]'
+        element :required_qualifications, '[data-qa=course__qualifications]'
+        element :personal_qualities, '[data-qa=course__personal_qualities]'
+        element :other_requirements, '[data-qa=course__other_requirements]'
+        element :has_vacancies, '[data-qa=course__has_vacancies]'
+        element :is_findable, '[data-qa=course__is_findable]'
+        element :open_for_applications, '[data-qa=course__open_for_applications]'
+        element :last_published_at, '[data-qa=course__last_published_date]'
+        element :status_tag, '[data-qa=course__content-status]'
+        element :preview_link, '[data-qa=course__preview-link]'
+        element :publish, '[data-qa=course__publish]'
+        element :success_summary, '.govuk-success-summary'
+        element :error_summary, '.govuk-error-summary'
+        element :status_panel, '[data-qa=course__status_panel]'
       end
     end
   end
